consolidate web response validation
