+++ react/eventwizard/net/EnsureRequestSuccess.js
@@ -1,0 +1,8 @@
+function ensureRequestSuccess(response) {
+  if (!response.ok) {
+    throw new Error(response.statusText);
+  }
+  return response;
+}
+
+export default ensureRequestSuccess
